% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerFileHandler.R
\docType{methods}
\name{addFile,LCMSAlignerFileHandler-method}
\alias{addFile,LCMSAlignerFileHandler-method}
\title{Add a file to LCMSALignerFIleHandler}
\usage{
\S4method{addFile}{LCMSAlignerFileHandler}(object, path)
}
\arguments{
\item{path}{The path to the file ot be added}

\item{lafh}{An LCMSAligenrFileHandler object}
}
\value{
The filled LCMSFIleHAandler object
}
\description{
Add a file to LCMSALignerFIleHandler
}
\examples{
print('Examples ot be put here')
}
