% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{computeDivergence}
\alias{computeDivergence}
\title{Compute all the divergence values of a set of graph}
\usage{
computeDivergence(lar, neighbours, points, ref_point)
}
\arguments{
\item{lar}{The reference object of an LCMSAlignerModel}

\item{neighbours}{The index of the neighbouring points}

\item{points}{A three column with mz,rt,int}

\item{ref_point}{The reference point in whichi to comapre this shit}
}
\description{
Compute all the divergence values of a set of graph
}
