% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{findBestCandidate}
\alias{findBestCandidate}
\title{Fidning the best candidates amongs a set of neighbours}
\usage{
findBestCandidate(lar, neighbours, points, ref_point, threshold = 5)
}
\arguments{
\item{lar}{The reference object of an LCMSAlignerModel}

\item{neighbours}{The index of the neighbouring points}

\item{points}{The full set of points the data table}

\item{ref_point}{Te reference point to consider}

\item{threshold}{The threshold in divergence distance to csondier}
}
\value{

}
\description{
Fidning the best candidates amongs a set of neighbours
}
