% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZmineXMLManipulator-package.R
\docType{package}
\name{MZmineXMLManipulator-package}
\alias{MZmineXMLManipulator-package}
\title{Parsing and generation of XML MZmine configuration files}
\description{
Parsing and generation of XML MZmine configuration files, notably to JSON and form JSON
}
\details{
Parsing and generation of XML MZmine configuration files, notably to JSON and form JSON. Test
}
