% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineLCMSaligner-package.R
\docType{package}
\name{onlineLCMSaligner-package}
\alias{onlineLCMSaligner-package}
\title{Online alignment of LC-MS files}
\description{
Online alignment of LC-MS files
}
\details{
Online alignment of LC-MS files
}
