% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot the number of detected files}
\usage{
plotHistogram(lam)
}
\arguments{
\item{lam}{A LCMSAlignerModel object}
}
\value{
The ggplot graph
}
\description{
Plot the number of detected files
}
\examples{
print('Examples to be put here')
}
